Move DynamicFilterData from obsolete dynamic_filter.hpp to new tablefilter_internal_functions
