+++ src/include/duckdb/planner/filter/dynamic_filter.hpp
@@ -10,29 +10,10 @@
 #pragma once
 
 #include "duckdb/planner/table_filter.hpp"
-#include "duckdb/common/types/value.hpp"
-#include "duckdb/common/enums/expression_type.hpp"
-#include "duckdb/common/atomic.hpp"
-#include "duckdb/common/mutex.hpp"
 
 namespace duckdb {
 
-struct DynamicFilterData {
-	DynamicFilterData() : comparison_type(ExpressionType::INVALID) {
-	}
-	DynamicFilterData(ExpressionType comparison_type_p, Value constant_p);
-
-	mutex lock;
-	ExpressionType comparison_type;
-	Value constant;
-	atomic<bool> initialized = {false};
-
-	void SetValue(Value val);
-	void Reset();
-	static bool CompareValue(ExpressionType comparison_type, const Value &constant, const Value &value);
-	static FilterPropagateResult CheckStatistics(BaseStatistics &stats, ExpressionType comparison_type,
-	                                             const Value &constant);
-};
+struct DynamicFilterData;
 
 //! DEPRECATED - only preserved for backwards-compatible deserialization and expression conversion
 class DynamicFilter : public TableFilter {
